First pass at simple sync point database handler completed

+++ branches/libindi-align/libs/indibase/alignment/AlignmentSubsystem.h
@@ -560,7 +560,7 @@ private:
     // Data
     struct DatabaseEntry
     {
-        double ObservationDate;
+        int ObservationDate;
         double ObservationTime;
         double RightAscension;
         double Declination;
